Moved logic to screeningUtils.js
